+++ inst/explorer/global/function_visualise_corpus.R
@@ -6,6 +6,8 @@ source("./global/corpus_plot_functions/label_axes.R", local = TRUE)
 
 source("./global/corpus_plot_functions/convenience_functions_corpus_vis.R", local = TRUE)
 
+source("./global/corpus_plot_functions/improve_visualisation_of_2_terms.R", local = TRUE)
+
 source("./global/corpus_plot_functions/create_coordinates_1_data_365.R", local = TRUE)
 source("./global/corpus_plot_functions/create_distance_coordinates_365.R", local = TRUE)
 source("./global/corpus_plot_functions/plotting_corpus_data_365.R", local = TRUE)
@@ -76,7 +78,7 @@ if (INCLUDE_EXTRA == TRUE) {
                                          df,
                                          search_arguments$case_sensitive,
                                          modus,
-                                         indices_included = doc_df$ID)
+                                         indices_included = doc_df$cx_ID)
       }
       colnames(count_overview) <- sprintf("Term_%s", seq_len(linjer))
   }